move package partItem
